+++ elftools/dwarf/structs.py
@@ -8,16 +8,15 @@
 # This code is in the public domain
 #-------------------------------------------------------------------------------
 from logging.config import valid_ident
-from elftools.construct.core import Subconstruct
-from elftools.construct.macros import Embedded, StreamOffset
 from ..construct import (
     UBInt8, UBInt16, UBInt32, UBInt64, ULInt8, ULInt16, ULInt32, ULInt64,
     SBInt8, SBInt16, SBInt32, SBInt64, SLInt8, SLInt16, SLInt32, SLInt64,
     Adapter, Struct, ConstructError, If, Enum, Array, PrefixedArray,
     CString, Embed, StaticField, IfThenElse, Construct, Rename, Sequence,
-    Switch, Value
+    Switch
     )
-from ..common.construct_utils import RepeatUntilExcluding, ULEB128, SLEB128
+from ..common.construct_utils import (RepeatUntilExcluding, ULEB128, SLEB128,
+    StreamOffset)
 from .enums import *
